% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BernoulliErrorChecks.R
\name{bernoulli_error_checks}
\alias{bernoulli_error_checks}
\title{Checks for Errors when Outcome is Bernoulli.}
\usage{
bernoulli_error_checks(effect_vals, hist_control_data, rand_control_diff,
  alpha)
}
\arguments{
\item{effect_vals}{See help page for \code{historic_sim()}.}

\item{hist_control_data}{See help page for \code{historic_sim()}.}

\item{rand_control_diff}{See help page for \code{historic_sim()}.}

\item{alpha}{See help page for \code{historic_sim()}.}
}
\value{
\code{bernoulli_error_checks()} returns messages when
\code{historic_sim()} function inputs are incorrectly specified.
Not to be called directly by user.
}
\description{
\code{bernoulli_error_checks()} function used only used internally by
\code{historic_sim()} function to check for proper input.  Not to be called
directly by user.
}
\examples{
#None
}
\keyword{internal}
