% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BernoulliCode.R
\name{bernoulliloglike}
\alias{bernoulliloglike}
\title{Log-likelihood function for two-arm trial with historical data using Bernoulli
distribution.}
\usage{
bernoulliloglike(params, randdata, histdata, a0)
}
\arguments{
\item{params}{Two element vecter of Bernoulli parameters.  The first and second elements
are the intercept (beta0) and treatment effect parameter (beta1), where the treatment effect is
a log odds ratio (experimental group over control group).  The prob parameter required by
dbinom() is equal to exp(params[1] + params[2]*treatment) / (1 + exp(params[1] + params[2]*treatment)).
It is assumed that the params[1] parameter is the same in both randomized and historical data.
It is assumed that the prob parameter for dbinom() in the randomized and historical control data is
equal to exp(params[1]) / (1 + exp(params[1])).}

\item{randdata}{Dataset of randomly generated trial data.  Randomized trial datasets
must have 3 columns: id, treatment, and y.  The value of treatment must be 0 (control)
or 1 (experimental).  The values of y must be 0 or 1.}

\item{histdata}{Dataset of historical data.  Historical datasets must have 3 columns: id,
treatment, and y.  The value of treatment should be 0.  The values of y must be 0 or 1.}

\item{a0}{Power prior parameter: 0 implies historical data is ignored and 1 implies
all information in historical data is used.}
}
\value{
\code{bernoulliloglike()} returns a value of the loglikelihood function
  given a set of Bernoulli parameters, randomly generated trial data, and observed
  historical data.
}
\description{
\code{bernoulliloglike()} function only used internally by
\code{bernoullitrialsimulator()} function to estimate Bernoulli model parameters
when clinical trial involves experimental and control groups as well as historical
control data.  The Bernoulli log-likelihood is calculated by modelling \code{data}
as a Bernoulli random variable. Not to be called directly by user.
}
\examples{
#None
}
\keyword{internal}
