% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BernoulliCode.R
\name{bernoullitrialsimulatornohist}
\alias{bernoullitrialsimulatornohist}
\title{Simulate a single randomized trial using a Bernoulli outcome but not including any information from
historical controls (Logistic regression model).}
\usage{
bernoullitrialsimulatornohist(sample_size_val, prob1_val, odds_ratio_val,
  alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{prob1_val}{prob parameter value for randomized control arm.  Used in call to \code{rbinom()}.}

\item{odds_ratio_val}{Desired odds ratio between randomized experimental and control groups.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{bernoullitrialsimulatornohist()} returns a vector of simulation results. The
  first element is an estimated odds ratio, the second element is the estimated
  variance of the log odds ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{bernoullitrialsimulator()} function only used internally by
\code{simple_bernoulli_sim()} function to run a single trial simulation involving historical
control data and a Bernoulli (0/1) outcome.
}
\details{
The simulation of a trial with a Bernoulli outcome without historical control data returns
an estimate of the odds ratio as well as an estimate of the log odds ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{bernoullitrialsimulatornohist()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
