% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianCode.R
\name{gaussianloglike}
\alias{gaussianloglike}
\title{Log-likelihood function for two-arm trial with historical data using Gaussian
distribution.}
\usage{
gaussianloglike(params, randdata, histdata, a0)
}
\arguments{
\item{params}{Three element vecter of Gaussian parameters.  Third element is log(sd),
where sd is a parameter required by dnorm().  The first and second elements
are the intercept and treatment effect parameter, where the treatment effect is
a mean difference (experimental group minus control group).  The mu parameter required by
dnorm() is equal to params[1] + params[2]*treatment.  It is assumed that the log(sd)
parameter is the same in both randomized and historical data.  It is assumed that
the mu parameter in the randomized and historical control data is equal to params[1].}

\item{randdata}{Dataset of randomly generated trial data.  Randomized trial datasets
must have 3 columns: id, treatment, and y.  The value of treatment must be 0 (control)
or 1 (experimental).  The values of y must be numeric.}

\item{histdata}{Dataset of historical data.  Historical datasets must have 3 columns: id,
treatment, and y.  The value of treatment should be 0.  The values of y must be
numeric.}

\item{a0}{Power prior parameter: 0 implies historical data is ignored and 1 implies
all information in historical data is used.}
}
\value{
\code{gaussianloglike()} returns a value of the loglikelihood function
  given a set of Gaussian parameters, randomly generated trial data, and observed
  historical data.
}
\description{
\code{gaussianloglike()} function used only used internally by
\code{gaussiantrialsimulator()} function to estimate Gaussian model parameters
when clinical trial involves experimental and control groups as well as historical
control data.  The Gaussian log-likelihood is calculated by modelling \code{data}
as a Gaussian random variable. Not to be called directly by user.
}
\examples{
#None
}
\keyword{internal}
