% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianCode.R
\name{gaussiantrialsimulatornohist}
\alias{gaussiantrialsimulatornohist}
\title{Simulate a single randomized trial using a Gaussian outcome but not including any information from
historical controls.}
\usage{
gaussiantrialsimulatornohist(sample_size_val, mu1_val, mean_diff_val,
  common_sd_val, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{mu1_val}{Randomized control arm mean parameter used in call to \code{rnorm()}.}

\item{mean_diff_val}{Desired mean difference between randomized experimental and control groups.}

\item{common_sd_val}{Randomized sd parameter used in call to \code{rnorm()}.
Used in both randomized arms.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{gaussiantrialsimulatornohist()} returns a vector of simulation results. The
  first element is an estimated mean difference, the second element is the estimated
  variance of the mean difference, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{gaussiantrialsimulatornohist()} function only used internally by
\code{simple_gaussian_sim()} function to estimate Gaussian model parameters
when clinical trial involves experimental and control groups but no historical control
data.
}
\details{
The simulation of a trial with a Gaussian outcome involving no historical control data returns
an estimate of the mean difference as well as an estimate of the mean difference variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{gaussiantrialsimulatornohist()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
