% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalErrorChecks.R
\name{global_error_checks}
\alias{global_error_checks}
\title{Checks \code{historic_sim()} Function Call for Global Type Errors.}
\usage{
global_error_checks(outcome_type, a0_vals, subj_per_arm, hist_control_data,
  rand_control_diff, get_var, get_bias, get_mse)
}
\arguments{
\item{outcome_type}{See help page for \code{historic_sim()}.}

\item{a0_vals}{See help page for \code{historic_sim()}.}

\item{subj_per_arm}{See help page for \code{historic_sim()}.}

\item{hist_control_data}{See help page for \code{historic_sim()}.}

\item{rand_control_diff}{See help page for \code{historic_sim()}.}

\item{get_var}{See help page for \code{historic_sim()}.}

\item{get_bias}{See help page for \code{historic_sim()}.}

\item{get_mse}{See help page for \code{historic_sim()}.}
}
\value{
\code{global_error_checks()} returns messages when \code{historic_sim()} function
inputs are incorrectly specified.  Not to be called directly by user.
}
\description{
\code{global_error_checks()} function used only used internally by \code{historic_sim()}
function to check for proper input.  Not to be called directly by user.
}
\examples{
#None
}
\keyword{internal}
