% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalErrorChecks.R
\name{global_error_checks_simple}
\alias{global_error_checks_simple}
\title{Checks \code{simple_sim()} Function Call for Global Type Errors.}
\usage{
global_error_checks_simple(outcome_type, subj_per_arm, get_var, get_bias,
  get_mse)
}
\arguments{
\item{outcome_type}{See help page for \code{simple_sim()}.}

\item{subj_per_arm}{See help page for \code{simple_sim()}.}

\item{get_var}{See help page for \code{simple_sim()}.}

\item{get_bias}{See help page for \code{simple_sim()}.}

\item{get_mse}{See help page for \code{simple_sim()}.}
}
\value{
\code{global_error_checks_simple()} returns messages when
\code{simple_sim()} function inputs are incorrectly specified.  Not to be called
directly by user.
}
\description{
\code{global_error_checks_simple()} function used only used internally by
\code{simple_sim()} function to check for proper input.  Not to be called
directly by user.
}
\examples{
#None
}
\keyword{internal}
