% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormalCode.R
\name{lognormalloglikenohist}
\alias{lognormalloglikenohist}
\title{Log-likelihood function for two-arm trial with no historical data using Lognormal distribution.}
\usage{
lognormalloglikenohist(params, randdata)
}
\arguments{
\item{params}{Three element vector of Lognormal parameters.  Third element is log(sd),
where sd is a parameter required by dnorm().  The first and second elements
are the intercept (beta0) and treatment effect parameter (beta1), where the treatment effect is
a log mean ratio (experimental group over control group).  The mu parameter required by
dnorm() is equal to params[1] + params[2]*treatment.  It is assumed that
the mu parameter in the randomized control data is equal to params[1].}

\item{randdata}{Dataset of randomly generated trial data.  Randomized trial datasets
must have 4 columns: id, treatment, event_time, and status.  The value of treatment
must be 0 (control) or 1 (experimental).  The values of event_time must be positive.
The values of status must be 0 (right censored event) or 1 (observed event).}
}
\value{
\code{lognormalloglikenohist()} returns a value of the loglikelihood function
  given a set of Lognormal parameters and randomly generated trial data.
}
\description{
\code{lognormalloglikenohist()} function only used internally by
\code{lognormaltrialsimulatornohist()} function to estimate Lognormal model parameters
when clinical trial involves experimental and control groups but no historical control
data.  The lognormal log-likelihood is calculated by modelling \code{log(data)} as
a Gaussian random variable. Not to be called directly by user.
}
\examples{
#None
}
\keyword{internal}
