% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormalCode.R
\name{lognormaltrialsimulatornohist}
\alias{lognormaltrialsimulatornohist}
\title{Simulate a single randomized trial using a Lognormal outcome but not including any information from
historical controls.}
\usage{
lognormaltrialsimulatornohist(sample_size_val, mu1_val, mean_ratio_val,
  common_sd_val, censor_value, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{mu1_val}{meanlog parameter used in call to \code{rlnorm()}.
Used only in randomized control arm.}

\item{mean_ratio_val}{Desired mean ratio between experimental and control arms.}

\item{common_sd_val}{sdlog parameter used in call to \code{rlnorm()}.
Used in both arms.}

\item{censor_value}{Value at which time-to-event data are right censored.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{lognormaltrialsimulatornohist()} returns a vector of simulation results. The
  first element is an estimated mean ratio, the second element is the estimated
  variance of the log mean ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{lognormaltrialsimulatornohist()} function only used internally by
\code{simple_lognormal_sim()} function to estimate Lognormal model parameters
when clinical trial involves experimental and control groups but no historical control
data.
}
\details{
The simulation of a trial with a Lognormal outcome involving no historical control data returns
an estimate of the mean ratio as well as an estimate of the log mean ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{lognormaltrialsimulatornohist()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
