% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalErrorChecks.R
\name{plot_table_error_checks}
\alias{plot_table_error_checks}
\title{Checks \code{plot_table()} Function Call for Errors.}
\usage{
plot_table_error_checks(bayes_ctd_robj, measure, tab_type, smooth, plot_out,
  subj_per_arm_val, a0_val, effect_val, rand_control_diff_val, span)
}
\arguments{
\item{bayes_ctd_robj}{See help page for \code{plot_table()}.}

\item{measure}{See help page for \code{plot_table()}.}

\item{tab_type}{See help page for \code{plot_table()}.}

\item{smooth}{See help page for \code{plot_table()}.}

\item{plot_out}{See help page for \code{plot_table()}.}

\item{subj_per_arm_val}{See help page for \code{plot_table()}.}

\item{a0_val}{See help page for \code{print_table()}.}

\item{effect_val}{See help page for \code{plot_table()}.}

\item{rand_control_diff_val}{See help page for \code{plot_table()}.}

\item{span}{See help page for \code{plot_table()}.}
}
\value{
\code{plot_table_error_checks()} returns messages when
\code{plot_table()} function inputs are incorrectly specified.  Not to be called
directly by user.
}
\description{
\code{plot_table_error_checks()} function used only used internally by
\code{plot_table()} function to check for proper input.  Not to be called
directly by user.
}
\examples{
#None
}
\keyword{internal}
