% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonCode.R
\name{poissonloglike}
\alias{poissonloglike}
\title{Log-likelihood function for two-arm trial with historical data using Poisson
distribution.}
\usage{
poissonloglike(params, randdata, histdata, a0)
}
\arguments{
\item{params}{Two element vecter of Poisson parameters.  The first and second elements
are the intercept (beta0) and treatment effect parameter (beta1), where the treatment effect is
a log mean ratio (experimental group over control group).  The lambda parameter required by
dpois() is equal to exp(params[1] + params[2]*treatment).  It is assumed that the params[1]
parameter is the same in both randomized and historical data.  It is assumed that
the lambda parameter in the randomized and historical control data is equal to exp(params[1]).}

\item{randdata}{Dataset of randomly generated trial data.  Randomized trial datasets
must have 3 columns: id, treatment, and y.  The value of treatment must be 0 (control)
or 1 (experimental).  The values of y must be non negative.}

\item{histdata}{Dataset of historical data.  Historical datasets must have 3 columns: id,
treatment, and y.  The value of treatment should be 0.  The values of y must be
non negative.}

\item{a0}{Power prior parameter: 0 implies historical data is ignored and 1 implies
all information in historical data is used.}
}
\value{
\code{poissonloglike()} returns a value of the loglikelihood function
  given a set of Poisson parameters, randomly generated trial data, and observed
  historical data.
}
\description{
\code{poissonloglike()} function only used internally by
\code{poissontrialsimulator()} function to estimate Poisson model parameters
when clinical trial involves experimental and control groups as well as historical
control data.  The Poisson log-likelihood is calculated by modelling \code{data}
as a Poisson random variable. Not to be called directly by user.
}
\examples{
#None
}
\keyword{internal}
