% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonCode.R
\name{poissontrialsimulator}
\alias{poissontrialsimulator}
\title{Simulate a single randomized trial using a Gaussian outcome and information from
historical controls.}
\usage{
poissontrialsimulator(sample_size_val, histdata, mu1_val, mean_ratio_val,
  a0_val, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{histdata}{Dataset of historical data.  Historical datasets must have 3 columns: id,
treatment, and y.  The value of treatment should be 0.  The values of y must be
non negative.}

\item{mu1_val}{Mean parameter value for randomized control arm.  Used in call to \code{rpois()}.}

\item{mean_ratio_val}{Desired mean ratio between randomized experimental and control groups.}

\item{a0_val}{A power prior parameter ranging from 0 to 1, where 0
implies no information from historical data should be used, 1 implies all of
the information from historical data should be used.  A value between 0 and 1
implies that a proportion of the information from historical data will be used.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{poissontrialsimulator()} returns a vector of simulation results. The
  first element is an estimated mean ratio, the second element is the estimated
  variance of the log mean ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{poissontrialsimulator()} function only used internally by
\code{poisson_sim()} function to run a single trial simulation involving historical
control data and a Poisson outcome.
}
\details{
The simulation of a trial with a Poisson outcome involving historical control data returns
an estimate of the mean ratio as well as an estimate of the log mean ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{poissontrialsimulator()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
