% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonCode.R
\name{poissontrialsimulatornohist}
\alias{poissontrialsimulatornohist}
\title{Simulate a single randomized trial using a Poisson outcome but not including any information from
historical controls.}
\usage{
poissontrialsimulatornohist(sample_size_val, mu1_val, mean_ratio_val, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{mu1_val}{Mean parameter value for randomized control arm.  Used in call to \code{rpois()}.}

\item{mean_ratio_val}{Desired mean ratio between randomized experimental and control groups.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{poissontrialsimulatornohist} returns a vector of simulation results. The
  first element is an estimated mean ratio, the second element is the estimated
  variance of the log mean ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{poissontrialsimulatornohist()} function only used internally by
\code{simple_poisson_sim()} function to estimate Poisson model parameters
when clinical trial involves experimental and control groups but no historical control
data.
}
\details{
The simulation of a trial with a Poisson outcome involving no historical control data returns
an estimate of the mean ratio as well as an estimate of the log mean ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{poissontrialsimulatornohist()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
