% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalErrorChecks.R
\name{print_table_error_checks}
\alias{print_table_error_checks}
\title{Checks \code{print_table()} Function Call for Errors.}
\usage{
print_table_error_checks(bayes_ctd_robj, measure, tab_type,
  subj_per_arm_val, a0_val, effect_val, rand_control_diff_val)
}
\arguments{
\item{bayes_ctd_robj}{See help page for \code{print_table()}.}

\item{measure, }{See help page for \code{print_table()}.}

\item{tab_type, }{See help page for \code{print_table()}.}

\item{subj_per_arm_val}{See help page for \code{print_table()}.}

\item{a0_val}{See help page for \code{print_table()}.}

\item{effect_val}{See help page for \code{print_table()}.}

\item{rand_control_diff_val}{See help page for \code{print_table()}.}
}
\value{
\code{print_table_error_checks()} returns messages when
\code{print_table} function inputs are incorrectly specified.  Not to be called
directly by user.
}
\description{
\code{print_table_error_checks()} function used only used internally by
\code{print_table()} function to check for proper input.  Not to be called
directly by user.
}
\examples{
#None.
}
\keyword{internal}
