% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWEErrorChecks.R
\name{pwe_error_checks}
\alias{pwe_error_checks}
\title{Checks for Errors when Outcome is PieceWise Exponential.}
\usage{
pwe_error_checks(effect_vals, hist_control_data, rand_control_diff,
  time_vec, censor_value, alpha)
}
\arguments{
\item{effect_vals}{See help page for \code{historic_sim()}.}

\item{hist_control_data}{See help page for \code{historic_sim()}.}

\item{rand_control_diff}{See help page for \code{historic_sim()}.}

\item{time_vec}{See help page for \code{historic_sim()}.}

\item{censor_value}{See help page for \code{historic_sim()}.}

\item{alpha}{See help page for \code{historic_sim()}.}
}
\value{
\code{pwe_error_checks()} returns messages when
\code{historic_sim()} function inputs are incorrectly specified.
Not to be called directly by user.
}
\description{
\code{pwe_error_checks()} function used only used internally by
\code{historic_sim()} function to check for proper input.  Not to be called
directly by user.
}
\examples{
#None
}
\keyword{internal}
