% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWECode.R
\name{pwe_loglike}
\alias{pwe_loglike}
\title{Log-likelihood function for two-arm trial with historical data using Piece-wise
Exponential (pwe) distribution.}
\usage{
pwe_loglike(params, time_vec, randdata, hist_data, a0)
}
\arguments{
\item{params}{A vecter of p + 1 parameters.  The last element is
the log hazard ratio.  The first p elements are the estimated log hazards
within each section created by \code{time_vec}.  If \code{time_vec} has
m elements, then p = m + 1.}

\item{time_vec}{Vector of cut points used to divide time-scale into sections
within which hazard is constant.}

\item{randdata}{Dataset of randomly generated trial data.}

\item{hist_data}{Dataset of historical data.}

\item{a0}{Power prior parameter: 0 implies historical data is ignored and 1 implies
all information in historical data is used.}
}
\value{
\code{pwe_loglike()} returns a value of the loglikelihood function
  given a set of pwe parameters, randomly generated trial data, and observed
  historical data.
}
\description{
\code{pwe_loglike()} function used only used internally by
\code{pwetrialsimulator()} function to estimate pwe model parameters
when clinical trial involves experimental and control groups as well as historical
control data.  The pwe log-likelihood is calculated by using \code{eha::dpch()}.
pwe_loglike() should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
