% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWECode.R
\name{pwe_sim}
\alias{pwe_sim}
\title{Repeated Two Arm Bayesian Clinical Trial Simulation with Historical Data and
Piece-wise Exponential Outcome.}
\usage{
pwe_sim(trial_reps = 100, subj_per_arm, a0_vals, effect_vals,
  rand_control_diff, hist_control_data, time_vec_val, censor_value,
  alpha = 0.05, get_var = FALSE, get_bias = FALSE, get_mse = FALSE,
  quietly = TRUE)
}
\arguments{
\item{trial_reps}{Number of trials to replicate within each combination of
a0_val, subj_per_arm, effect_vals, and rand_control_diff.  As the number
of trials increases, the precision of the estimate will increase. Default is
100.}

\item{subj_per_arm}{A vector of sample sizes, all of which must be positive
integers.}

\item{a0_vals}{A vector of power prior parameters ranging from 0 to 1, where 0
implies no information from historical data should be used, 1 implies all of
the information from historical data should be used.  A value between 0 and 1
implies that a proportion of the information from historical data will be used.}

\item{effect_vals}{A vector of hazard ratios (randomized experimental over control),
all of which must be positive.}

\item{rand_control_diff}{For piece-wise exponential outcomes this is a vector of hazard ratios
(randomized controls over historical controls) that represent differences
between randomized and historical controls.}

\item{hist_control_data}{A dataset of historical data.  Default is \code{NULL}.
Historical datasets must have 4 columns: id, treatment, event_time, status, and
indicator.  The value of treatment should be 0.  The values of event_time must
be positive.  The values of status must be 0 (right censored event) or
1 (observed event).  The value of indicator must be 2.}

\item{time_vec_val}{A vector of time values which are used to create time periods
within which the exponential hazard is constant.  Only used for piecewise
exponential models.  Default is \code{NULL}.}

\item{censor_value}{A single value at which right censoring occurs when
simulating randomized subject outcomes.  Default is \code{NULL}, where
\code{NULL} implies no right censoring.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}

\item{get_var}{A TRUE/FALSE indicator of whether an array of variance
estimates will be returned.  Default is \code{FALSE}.}

\item{get_bias}{A TRUE/FALSE indicator of whether an array of bias
estimates will be returned.  Default is \code{FALSE}.}

\item{get_mse}{A TRUE/FALSE indicator of whether an array of MSE
estimates will be returned.  Default is \code{FALSE}.}

\item{quietly}{A TRUE/FALSE indicator of whether notes are printed
to output about simulation progress as the simulation runs.  If
running interactively in RStudio or running in the R console,
\code{quietly} can be set to FALSE.  If running in a Notebook or
knitr document, \code{quietly} needs to be set to TRUE.  Otherwise
each note will be printed on a separate line and it will take up
a lot of output space.  Default is \code{TRUE}.}
}
\value{
\code{pwe_sim()} returns an S3 object of class bayes_ctd_array.
}
\description{
\code{pwe_sim()} function only used internally by \code{historic_sim()}
function to run a set of trial simulations involving historical
control data and a piece-wise exponential outcome.  User defined simulation parameters are
used to generate a set of trial scenarios.  Each scenario is simulated multiple
times and then means are taken to calculate estimates of power, hazard ratios,
and other user requested study summary statistics like variance of hazard
ratio, bias (on hazard ratio scale), and mse (on hazard ratio scale).
The number of repeated simulations is defined by the user.
}
\details{
\code{pwe_sim()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
