% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWECode.R
\name{pwetrialsimulator}
\alias{pwetrialsimulator}
\title{Simulate a single randomized trial using a piece-wise exponential outcome and information from
historical controls.}
\usage{
pwetrialsimulator(sample_size_val, hist_data, lambda_vec_val, time_vec_val,
  hazard_ratio_val, censor_value, a0_val, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{hist_data}{Dataset of historical data.  Historical datasets must have 5 columns:
id, treatment, event_time, status, and indicator.  The value of treatment should be 0.  The
values of event_time must be positive.  The values of status must be 0 (right
censored event) or 1 (observed event).  The value of indicator must be 2.}

\item{lambda_vec_val}{Set of lambdas for the randomized control arm that will be
passed to \code{eha::rpch()} through the levels parameter.}

\item{time_vec_val}{Set of cutpoints for the simulated trial data that will be
passed to \code{eha::rpch()} through the cuts parameter.}

\item{hazard_ratio_val}{Desired Hazard Ratio between randomized experimental and control groups.}

\item{censor_value}{Value at which time-to-event data are right censored.}

\item{a0_val}{A power prior parameter ranging from 0 to 1, where 0
implies no information from historical data should be used, 1 implies all of
the information from historical data should be used.  A value between 0 and 1
implies that a proportion of the information from historical data will be used.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{pwetrialsimulator()} returns a vector of simulation results. The
  first element is an estimated hazard ratio, the second element is the estimated
  variance of the log hazard ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{pwetrialsimulator()} function only used internally by
\code{pwe_sim()} function to run a single trial simulation involving historical
control data and a piece-wise exponential outcome.
}
\details{
The simulation of a trial with a piece-wise exponential outcome involving historical control data returns
an estimate of the hazard ratio as well as an estimate of the log hazard ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{pwetrialsimulator()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
