% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWECode.R
\name{pwetrialsimulatornohist}
\alias{pwetrialsimulatornohist}
\title{Simulate a single randomized trial using a piece-wise exponential outcome but not including any
information from historical controls.}
\usage{
pwetrialsimulatornohist(sample_size_val, lambda_vec_val, time_vec_val,
  hazard_ratio_val, censor_value, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{lambda_vec_val}{Set of lambdas for the randomized control arm that will be
passed to \code{eha::rpch()} through the levels parameter.}

\item{time_vec_val}{Set of cutpoints for the simulated trial data that will be
passed to \code{eha::rpch()} through the cuts parameter.}

\item{hazard_ratio_val}{Desired Hazard Ratio between randomized experimental and control groups.}

\item{censor_value}{Value at which time-to-event data are right censored.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{pwetrialsimulatornohist()} returns a vector of simulation results. The
  first element is an estimated hazard ratio, the second element is the estimated
  variance of the log hazard ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{pwetrialsimulatornohist()} function only used internally by
\code{simple_pwe_sim()} function to estimate piece-wise exponential model parameters
when clinical trial involves experimental and control groups but no historical control
data.
}
\details{
The simulation of a trial with a piece-wise exponential outcome involving no historical control
data returns an estimate of the hazard ratio as well as an estimate of the log hazard ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{pwetrialsimulatornohist()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
