% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BernoulliCode.R
\name{simple_bernoulli_sim}
\alias{simple_bernoulli_sim}
\title{Repeated Two Arm Bayesian Clinical Trial Simulation with no Historical Data and
Bernoulli Outcome (Logistic regression model).}
\usage{
simple_bernoulli_sim(trial_reps, subj_per_arm, effect_vals, prob1_val, alpha,
  get_var, get_bias, get_mse)
}
\arguments{
\item{trial_reps}{Number of trials to replicate within each combination of
subj_per_arm and effect_vals.  As the number of trials increases, the
precision of the estimate will increase. Default is 100.}

\item{subj_per_arm}{A vector of sample sizes, all of which must be positve
integers.}

\item{effect_vals}{A vector of odds ratios (randomized experimental over control),
all of which must be positive.}

\item{prob1_val}{prob parameter value for randomized control arm. Used in call to \code{rbinom()}.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}

\item{get_var}{A TRUE/FALSE indicator of whether or not an array of variance
estimates will be returned.  Default is \code{FALSE}.}

\item{get_bias}{A TRUE/FALSE indicator of whether or not an array of bias
estimates will be returned.  Default is \code{FALSE}.}

\item{get_mse}{A TRUE/FALSE indicator of whether or not an array of MSE
estimates will be returned.  Default is \code{FALSE}.}
}
\value{
\code{simple_bernoulli_sim()} returns an S3 object of class bayes_ctd_array.
}
\description{
\code{simple_bernoulli_sim()} function only used internally by \code{simple_sim()}
function to run a set of trial simulations involving no historical
control data and a Bernoulli outcome.  User defined simulation parameters are
used to generate a set of trial scenarios.  Each scenario is simulated multiple
times and then means are taken to calculate estimates of power, odds ratios,
and other user requested study summary statistics like variance of odds
ratio, bias (on odds ratio scale), and mse (on odds ratio scale).
The number of repeated simulations is defined by the user.
}
\details{
\code{simple_bernoulli_sim()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
