% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonCode.R
\name{simple_poisson_sim}
\alias{simple_poisson_sim}
\title{Repeated Two Arm Bayesian Clinical Trial Simulation with no Historical Data and
Poisson Outcome.}
\usage{
simple_poisson_sim(trial_reps, subj_per_arm, effect_vals, mu1_val, alpha,
  get_var, get_bias, get_mse)
}
\arguments{
\item{trial_reps}{Number of trials to replicate within each combination of
subj_per_arm and effect_vals.  As the number of trials increases, the
precision of the estimate will increase. Default is 100.}

\item{subj_per_arm}{A vector of sample sizes, all of which must be positve
integers.}

\item{effect_vals}{A vector of mean ratios (randomized experimental over control),
all of which must be positive.}

\item{mu1_val}{lambda parameter value for randomized control arm. Used in call to \code{rpois()}.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}

\item{get_var}{A TRUE/FALSE indicator of whether an array of variance
estimates will be returned.  Default is \code{FALSE}.}

\item{get_bias}{A TRUE/FALSE indicator of whether an array of bias
estimates will be returned.  Default is \code{FALSE}.}

\item{get_mse}{A TRUE/FALSE indicator of whether an array of MSE
estimates will be returned.  Default is \code{FALSE}.}
}
\value{
\code{simple_poisson_sim()} returns an S3 object of class bayes_ctd_array.
}
\description{
\code{simple_poisson_sim()} function only used internally by \code{simple_sim()}
function to run a set of trial simulations involving no historical
control data and a Poisson outcome.  User defined simulation parameters are
used to generate a set of trial scenarios.  Each scenario is simulated multiple
times and then means are taken to calculate estimates of power, mean ratios,
and other user requested study summary statistics like variance of mean
ratio, bias (on mean ratio scale), and mse (on mean ratio scale).
The number of repeated simulations is defined by the user.
}
\details{
\code{simple_poisson_sim()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
