% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWECode.R
\name{simple_pwe_sim}
\alias{simple_pwe_sim}
\title{Repeated Two Arm Bayesian Clinical Trial Simulation with no Historical Data and
Piece-wise Exponential Outcome.}
\usage{
simple_pwe_sim(trial_reps, subj_per_arm, effect_vals, time_vec_val, rc_hazards,
  censor_value, alpha, get_var, get_bias, get_mse)
}
\arguments{
\item{trial_reps}{Number of trials to replicate within each combination of
a0_val, subj_per_arm, effect_vals, and rand_control_diff.  As the number
of trials increases, the precision of the estimate will increase. Default is
100.}

\item{subj_per_arm}{A vector of sample sizes, all of which must be positve
integers.}

\item{effect_vals}{A vector of hazard ratios (randomized experimental over control),
all of which must be positive.}

\item{time_vec_val}{Set of cutpoints for the simulated trial data that will be
passed to \code{eha::rpch()} through the cuts parameter.}

\item{rc_hazards}{Set of lambdas for the randomized control arm that will be
passed to \code{eha::rpch()} through the levels parameter.}

\item{censor_value}{A single value at which right censoring occurs when
simulating randomized subject outcomes.  Default is \code{NULL}, where
\code{NULL} implies no right censoring.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}

\item{get_var}{A TRUE/FALSE indicator of whether an array of variance
estimates will be returned.  Default is \code{FALSE}.}

\item{get_bias}{A TRUE/FALSE indicator of whether an array of bias
estimates will be returned.  Default is \code{FALSE}.}

\item{get_mse}{A TRUE/FALSE indicator of whether an array of MSE
estimates will be returned.  Default is \code{FALSE}.}
}
\value{
\code{simple_pwe_sim()} returns an S3 object of class bayes_ctd_array.
}
\description{
\code{simple_pwe_sim()} function only used internally by \code{simple_sim()}
function to run a set of trial simulations involving no historical
control data and a piece-wise exponential outcome.  User defined simulation parameters are
used to generate a set of trial scenarios.  Each scenario is simulated multiple
times and then means are taken to calculate estimates of power, hazard ratios,
and other user requested study summary statistics like variance of hazard
ratio, bias (on hazard ratio scale), and mse (on hazard ratio scale).
The number of repeated simulations is defined by the user.
}
\details{
\code{simple_pwe_sim()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
