% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullCode.R
\name{simple_weibull_sim}
\alias{simple_weibull_sim}
\title{Two Arm Bayesian Clinical Trial Simulation with no Historical Data and
Weibull Outcome.}
\usage{
simple_weibull_sim(trial_reps = 100, subj_per_arm, effect_vals,
  scale1_value, common_shape_value, censor_value, alpha = 0.05,
  get_var = FALSE, get_bias = FALSE, get_mse = FALSE,
  quietly = TRUE)
}
\arguments{
\item{trial_reps}{Number of trials to replicate within each combination of
subj_per_arm and effect_vals.  As the number of trials increases, the
precision of the estimate will increase. Default is 100.}

\item{subj_per_arm}{A vector of sample sizes, all of which must be positve
integers.}

\item{effect_vals}{A vector of hazard ratios (randomized experimental over control),
all of which must be positive.}

\item{scale1_value}{scale parameter value for randomized controls. Used in call
to \code{rweibull()}.}

\item{common_shape_value}{shape parameter value assumed common in both arms.
Used in call to \code{rweibull()}.}

\item{censor_value}{A single value at which right censoring occurs when
simulating randomized subject outcomes.  Default is \code{NULL}, where
\code{NULL} implies no right censoring.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}

\item{get_var}{A TRUE/FALSE indicator of whether an array of variance
estimates will be returned.  Default is \code{FALSE}.}

\item{get_bias}{A TRUE/FALSE indicator of whether an array of bias
estimates will be returned.  Default is \code{FALSE}.}

\item{get_mse}{A TRUE/FALSE indicator of whether an array of MSE
estimates will be returned.  Default is \code{FALSE}.}

\item{quietly}{A TRUE/FALSE indicator of whether notes are printed
to output about simulation progress as the simulation runs.  If
running interactively in RStudio or running in the R console,
\code{quietly} can be set to FALSE.  If running in a Notebook or
knitr document, \code{quietly} needs to be set to TRUE.  Otherwise
each note will be printed on a separate line and it will take up
alot of output space.  Default is \code{TRUE}.}
}
\value{
\code{simple_weibull_sim()} returns an S3 object of class bayes_ctd_array.
}
\description{
\code{simple_weibull_sim()} function only used internally by
\code{simple_sim()} function to run a set of trial simulations involving no
historical control data and a Weibull outcome.  User defined simulation
parameters are used to generate a set of trial scenarios.  Each scenario is
simulated multiple times and then means are taken to calculate estimates
of power, hazard ratios, and other user requested study summary statistics
like variance of hazard ratio, bias (on hazard ratrio scale), and
mse (on hazard ratrio scale).  The number of repeated simulations is
defined by the user.
}
\details{
\code{simple_weibull_sim()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
