% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullErrorChecks.R
\name{weibull_error_checks_simple}
\alias{weibull_error_checks_simple}
\title{Checks for Errors when Outcome is Weibull.}
\usage{
weibull_error_checks_simple(effect_vals, control_parms, censor_value,
  alpha)
}
\arguments{
\item{effect_vals}{See help page for \code{simple_sim()}.}

\item{control_parms}{See help page for \code{simple_sim()}.}

\item{censor_value}{See help page for \code{simple_sim()}.}

\item{alpha}{See help page for \code{simple_sim()}.}
}
\value{
\code{weibull_error_checks_simple()} returns messages when
\code{simple_sim()} function inputs are incorrectly specified.
Not to be called directly by user.
}
\description{
\code{weibull_error_checks_simple()} function used only used internally by
\code{simple_sim()} function to check for proper input.  Not to be called
directly by user.
}
\examples{
#None
}
\keyword{internal}
