% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullCode.R
\name{weibull_sim}
\alias{weibull_sim}
\title{Repeated Two Arm Bayesian Clinical Trial Simulation with Historical Data and
Weibull Outcome.}
\usage{
weibull_sim(trial_reps, subj_per_arm, a0_vals, effect_vals, rand_control_diff,
  hist_control_data, censor_value, alpha, get_var, get_bias, get_mse)
}
\arguments{
\item{trial_reps}{Number of trials to replicate within each combination of
a0_val, subj_per_arm, effect_vals, and rand_control_diff.  As the number
of trials increases, the precision of the estimate will increase. Default is
100.}

\item{subj_per_arm}{A vector of sample sizes, all of which must be positve
integers.}

\item{a0_vals}{A vector of power prior parameters ranging from 0 to 1, where 0
implies no information from historical data should be used, 1 implies all of
the information from historical data should be used.  A value between 0 and 1
implies that a proportion of the information from historical data will be used.}

\item{effect_vals}{A vector of hazard ratios between randomized arms (randomized
experimental over control), all of which must be positive.}

\item{rand_control_diff}{A vector of hazard ratios (randomized controls over
historical controls) representing differences between historical and randomized
controls.}

\item{hist_control_data}{A datset of historical data.  Default is \code{NULL}.
Historical datasets must have 4 columns: id, treatment, event_time, and
status.  The value of treatment should be 0.  The values of event_time must
be positive.  The values of status must be 0 (right censored event) or
1 (observed event).}

\item{censor_value}{A single value at which right censoring occurs when
simulating randomized subject outcomes.  Default is \code{NULL}, where
\code{NULL} implies no right censoring.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}

\item{get_var}{A TRUE/FALSE indicator of whether an array of variance
estimates will be returned.  Default is \code{FALSE}.}

\item{get_bias}{A TRUE/FALSE indicator of whether an array of bias
estimates will be returned.  Default is \code{FALSE}.}

\item{get_mse}{A TRUE/FALSE indicator of whether an array of MSE
estimates will be returned.  Default is \code{FALSE}.}
}
\value{
\code{weibull_sim()} returns an S3 object of class bayes_ctd_array.
}
\description{
\code{weibull_sim()} function only used internally by \code{historic_sim()}
function to run a set of trial simulations involving historical
control data and a Weibull outcome.  User defined simulation parameters are
used to generate a set of trial scenarios.  Each scenario is simulated multiple
times and then means are taken to calculate estimates of power, hazard ratios,
and other user requested study summary statistics like variance of hazard
ratio, bias (on hazard ratrio scale), and mse (on hazard ratrio scale).
The number of repeated simulations is defined by the user.
}
\details{
\code{weibull_sim()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
