% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullCode.R
\name{weibullloglike}
\alias{weibullloglike}
\title{Log-likelihood function for two-arm trial with historical data using Weibull distribution.}
\usage{
weibullloglike(params, randdata, histdata, a0)
}
\arguments{
\item{params}{Three element vecter of Weibull parameters.  The third element is
the shape parameter used in \code{dweibull()}.  The first and second elements
are the intercept (beta0), and treatment effect (beta1), parameters as defined in
details section.  The beta1 parameter is the log hazard ratio.}

\item{randdata}{Dataset of randomly generated trial data.  Randomized trial datasets
must have 4 columns: id, treatment, event_time, and status.  The value of treatment
must be 0 (control) or 1 (experimental).  The values of event_time must be positive.
The values of status must be 0 (right censored event) or 1 (observed event).}

\item{histdata}{Dataset of historical control data.  Historical datasets must have 4 columns:
id, treatment, event_time, and status.  The value of treatment should be 0.  The
values of event_time must be positive.  The values of status must be 0 (right
censored event) or 1 (observed event).}

\item{a0}{Power prior parameter where 0 implies historical control data is ignored and 1 implies
all information in historical control data is used.  A value between 0 and 1 partially includes
the historical control data.}
}
\value{
\code{weibullloglike()} returns a value of the loglikelihood function
  given a set of Weibull parameters, randomly generated trial data, and observed
  historical control data.
}
\description{
\code{weibullloglike()} function only used internally by
\code{weibulltrialsimulator()} function to estimate Weibull model parameters
when clinical trial involves experimental and control groups as well as historical
control data.
}
\details{
The Weibull log-likelihood is calculated by using \code{dweibull()}.  The following
derivation was used to reparameterize the model parameters so that \code{dweibull()}
could be used in the log-likelihood function and still estimate the hazard ratio as a
parameter.

If we let \eqn{scale = exp(-beta0/shape - beta1*treatment/shape)}, then that for the control
group: \eqn{scale0 = exp(-beta0/shape)}.  Similarly, for the experimental group
\eqn{scale1 = exp(-beta0/shape - beta1/shape)}.

Now for the Weibull distribution \deqn{hazard(t) = (shape)*(scale^(-shape))*(t^(shape-1))}
From this equation, we can derive the Weibull hazard ratio, HR, (experimental over control) as a
function of \eqn{scale0}, \eqn{scale1}, and the shape parameter \eqn{HR = (scale0/scale1)^shape}.

Substituting for \eqn{scale0} and eqn{scale1} using functions of \eqn{beta0}, \eqn{beta1}, and
\eqn{shape}, we can espress the hazard ratio in terms of \eqn{beta0} and \eqn{beta1}. When
we do this we have \deqn{HR = (exp(-beta0/shape)/exp(-beta0/shape - beta1/shape))^shape}
After reducing terms we have a simply formula for the hazard ratio,
\eqn{HR = (exp(beta1/shape))^shape}, which reduces further to \eqn{exp(beta1)}.

Therefore, \eqn{beta1} is the log hazard ratio of experimental over control.  Similarly
\eqn{log(scale0) = -beta0/shape} so the nusance parameter \eqn{beta0} is equal to
\eqn{(-shape)log(scale0)}.

\code{weibullloglike()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
