% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullCode.R
\name{weibullloglikenohist}
\alias{weibullloglikenohist}
\title{Log-likelihood function for two-arm trial with no historical data using Weibull distribution.}
\usage{
weibullloglikenohist(params, randdata)
}
\arguments{
\item{params}{Three element vector of Weibull parameters.  The third element is
the shape parameter used in \code{dweibull()}.  The first and second elements
are the intercept (beta0), and treatment effect (beta1), parameters as defined in
details section.  The beta1 parameter is the log hazard ratio.}

\item{randdata}{Dataset of randomly generated trial data. Randomized trial datasets
must have 4 columns: id, treatment, event_time, and status.  The value of treatment
must be 0 (control) or 1 (experimental).  The values of event_time must be positive.
The values of status must be 0 (right censored event) or 1 (observed event).}
}
\value{
\code{weibullloglikenohist()} returns a value of the loglikelihood function
  given a set of Weibull parameters and randomly generated trial data.
}
\description{
\code{weibullloglikenohist()} function only used internally by
\code{weibulltrialsimulatornohist()} function to estimate Weibull model parameters
when clinical trial involves experimental and control groups but no historical control
data.
}
\details{
The Weibull log-likelihood is calculated by using \code{dweibull()}.  The following
derivation was used to reparametrize the model parameters so that \code{dweibull()}
could be used in the log-likelihood function and still estimate the hazard ratio as a
parameter.

If we let \eqn{scale = exp(-beta0/shape - beta1*treatment/shape)}, then that for the control
group: \eqn{scale0 = exp(-beta0/shape)}.  Similarly, for the experimental group
\eqn{scale1 = exp(-beta0/shape - beta1/shape)}.

Now for the Weibull distribution \deqn{hazard(t) = (shape)*(scale^(-shape))*(t^(shape-1))}
From this equation, we can derive the Weibull hazard ratio, HR, (experimental over control) as a
function of \eqn{scale0}, \eqn{scale1}, and the shape parameter \eqn{HR = (scale0/scale1)^shape}.

Substituting for \eqn{scale0} and eqn{scale1} using functions of \eqn{beta0}, \eqn{beta1}, and
\eqn{shape}, we can express the hazard ratio in terms of \eqn{beta0} and \eqn{beta1}. When
we do this we have \deqn{HR = (exp(-beta0/shape)/exp(-beta0/shape - beta1/shape))^shape}
After reducing terms we have a simply formula for the hazard ratio,
\eqn{HR = (exp(beta1/shape))^shape}, which reduces further to \eqn{exp(beta1)}.

Therefore, \eqn{beta1} is the log hazard ratio of experimental over control.  Similarly
\eqn{log(scale0) = -beta0/shape} so the nuisance parameter \eqn{beta0} is equal to
\eqn{(-shape)log(scale0)}.

\code{weibullloglike()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
