% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullCode.R
\name{weibulltrialsimulator}
\alias{weibulltrialsimulator}
\title{Simulate a single randomized trial using a Weibull outcome and information from
historical controls.}
\usage{
weibulltrialsimulator(sample_size_val, histdata, scale1_val,
  hazard_ratio_val, common_shape_val, censor_value, a0_val, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{histdata}{Dataset of historical data.  Historical datasets must have 4 columns:
id, treatment, event_time, and status.  The value of treatment should be 0.  The
values of event_time must be positive.  The values of status must be 0 (right
censored event) or 1 (observed event).}

\item{scale1_val}{Randomized control arm scale parameter used in call to \code{rweibull()}.}

\item{hazard_ratio_val}{Desired hazard ratio between randomized experimental and control arms.}

\item{common_shape_val}{Randomized shape parameter used in call to \code{rweibull()}.
Used in both randomized arms.}

\item{censor_value}{Value at which time-to-event data are right censored.}

\item{a0_val}{A power prior parameter ranging from 0 to 1, where 0
implies no information from historical data should be used, 1 implies all of
the information from historical data should be used.  A value between 0 and 1
implies that a proportion of the information from historical data will be used.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{weibulltrialsimulator()} returns a vector of simulation results. The
  first element is an estimated hazard ratio, the second element is the estimated
  variance of the log hazard ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{weibulltrialsimulator()} function only used internally by
\code{weibull_sim()} function to run a single trial simulation involving historical
control data and a Weibull outcome.
}
\details{
The simulation of a trial with a Weibull outcome involving historical control data returns
an estimate of the hazard ratio as well as an estimate of the log hazard ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{weibulltrialsimulator()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
