% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullCode.R
\name{weibulltrialsimulatornohist}
\alias{weibulltrialsimulatornohist}
\title{Simulate a single randomized trial using a Weibull outcome but not including any information from
historical controls.}
\usage{
weibulltrialsimulatornohist(sample_size_val, scale1_val, hazard_ratio_val,
  common_shape_val, censor_value, alpha)
}
\arguments{
\item{sample_size_val}{Number of subjects per arm.}

\item{scale1_val}{Scale parameter used in call to \code{rweibull()}.
Used only in randomized control arm.}

\item{hazard_ratio_val}{Desired hazard ratio between experimental and control arms.}

\item{common_shape_val}{Shape parameter used in call to \code{rweibull()}.
Used in both randomized arms.}

\item{censor_value}{Value at which time-to-event data are right censored.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}
}
\value{
\code{weibulltrialsimulatornohist()} returns a vector of simulation results. The
  first element is an estimated hazard ratio, the second element is the estimated
  variance of the log hazard ratio, and the third element is a 0/1 variable indicator
  whether or not the trial rejected the null hypothesis (1) or failed to reject
  the null hypothesis (0).
}
\description{
\code{weibulltrialsimulatornohist()} function only used internally by
\code{simple_weibull_sim()} function to run a single trial simulation involving
a Weibull outcome but no historical control data.
}
\details{
The simulation of a trial with a Weibull outcome involving no historical control data returns
an estimate of the hazard ratio as well as an estimate of the log hazard ratio variance.
Finally the simulation returns an indication of whether or not the simulated trial led to
a rejection of the null hypothesis (1) or not (0).

\code{weibulltrialsimulatornohist()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
