% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianCode.R
\name{simple_gaussian_sim}
\alias{simple_gaussian_sim}
\title{Repeated Two Arm Bayesian Clinical Trial Simulation with no Historical Data and
Gaussian Outcome.}
\usage{
simple_gaussian_sim(trial_reps = 100, subj_per_arm, effect_vals, mu1_val,
  common_sd_val, alpha = 0.05, get_var = FALSE, get_bias = FALSE,
  get_mse = FALSE, quietly = TRUE)
}
\arguments{
\item{trial_reps}{Number of trials to replicate within each combination of
subj_per_arm and effect_vals.  As the number of trials increases, the
precision of the estimate will increase. Default is 100.}

\item{subj_per_arm}{A vector of sample sizes, all of which must be positive
integers.}

\item{effect_vals}{A vector of mean differences (randomized experimental minus
controls).}

\item{mu1_val}{mean parameter value for randomized control arm. Used in call to
\code{rnorm()}.}

\item{common_sd_val}{sd parameter value used in both randomized arms.  Used in call
to \code{rnorm()}.}

\item{alpha}{A number ranging between 0 and 1 that defines the acceptable Type 1
error rate. Default is 0.05.}

\item{get_var}{A TRUE/FALSE indicator of whether an array of variance
estimates will be returned.  Default is \code{FALSE}.}

\item{get_bias}{A TRUE/FALSE indicator of whether an array of bias
estimates will be returned.  Default is \code{FALSE}.}

\item{get_mse}{A TRUE/FALSE indicator of whether an array of MSE
estimates will be returned.  Default is \code{FALSE}.}

\item{quietly}{A TRUE/FALSE indicator of whether notes are printed
to output about simulation progress as the simulation runs.  If
running interactively in RStudio or running in the R console,
\code{quietly} can be set to FALSE.  If running in a Notebook or
knitr document, \code{quietly} needs to be set to TRUE.  Otherwise
each note will be printed on a separate line and it will take up
a lot of output space.  Default is \code{TRUE}.}
}
\value{
\code{simple_gaussian_sim()} returns an S3 object of class bayes_ctd_array.
}
\description{
\code{simple_gaussian_sim()} function only used internally by \code{simple_sim()}
function to run a set of trial simulations involving no historical
control data and a Gaussian outcome.  User defined simulation parameters are
used to generate a set of trial scenarios.  Each scenario is simulated multiple
times and then means are taken to calculate estimates of power, mean difference,
and other user requested study summary statistics like variance of mean
difference, bias, and mse.  The number of repeated simulations is defined
by the user.
}
\details{
\code{simple_gaussian_sim()} should not be called directly by user.
}
\examples{
#None
}
\keyword{internal}
